Add Model*s submodules
