+++ test/test_belief_propagation.jl
@@ -14,7 +14,6 @@ using ITensorNetworks:
   contraction_sequence,
   environment,
   flatten_networks,
-  ising_network,
   linkinds_combiners,
   random_itensornetwork,
   siteinds,
@@ -22,6 +21,7 @@ using ITensorNetworks:
   tensornetwork,
   update,
   update_factor
+using ITensorNetworks.ModelNetworks: ModelNetworks
 using ITensors: ITensors, ITensor, combiner, dag, inds, op, prime, randomITensor
 using ITensors.NDTensors: array
 using LinearAlgebra: eigvals, tr
@@ -94,13 +94,12 @@ ITensors.disable_warn_order()
   s = IndsNetwork(g; link_space=2)
   beta = 0.2
   vs = [(2, 3), (3, 3)]
-  ψψ = ising_network(s, beta)
-  ψOψ = ising_network(s, beta; szverts=vs)
+  ψψ = ModelNetworks.ising_network(s, beta)
+  ψOψ = ModelNetworks.ising_network(s, beta; szverts=vs)
 
   contract_seq = contraction_sequence(ψψ)
   actual_szsz =
-    ITensors.contract(ψOψ; sequence=contract_seq)[] /
-    ITensors.contract(ψψ; sequence=contract_seq)[]
+    contract(ψOψ; sequence=contract_seq)[] / contract(ψψ; sequence=contract_seq)[]
 
   bpc = BeliefPropagationCache(ψψ, group(v -> v[1], vertices(ψψ)))
   bpc = update(bpc; maxiter=20)
@@ -125,9 +124,7 @@ ITensors.disable_warn_order()
 
   ψψsplit = split_index(ψψ, NamedEdge.([(v, 1) => (v, 2) for v in vs]))
   env_tensors = environment(bpc, [(v, 2) for v in vs])
-  rdm = ITensors.contract(
-    vcat(env_tensors, ITensor[ψψsplit[vp] for vp in [(v, 2) for v in vs]])
-  )
+  rdm = contract(vcat(env_tensors, ITensor[ψψsplit[vp] for vp in [(v, 2) for v in vs]]))
 
   rdm = array((rdm * combiner(inds(rdm; plev=0)...)) * combiner(inds(rdm; plev=1)...))
   rdm /= tr(rdm)
